Move out `Network` and `PendingWallet` from `wallet.rs`
