+++ rust/src/lib.rs
@@ -9,10 +9,12 @@ pub mod autocomplete;
 pub mod color_scheme;
 pub mod keychain;
 pub mod keys;
+pub mod mnemonic;
+pub mod network;
+pub mod pending_wallet;
 pub mod redb;
 pub mod view_model;
 pub mod wallet;
 pub mod word_validator;
-pub mod mnemonic;
 
 uniffi::setup_scaffolding!();
